Change long to long long to support big age

+++ C Primer Plus/Charpter 3/3.5.c
@@ -2,7 +2,7 @@
  * @Author: Tanaka Hao 314308943@qq.com
  * @Date: 2022-11-12 20:32:01
  * @LastEditors: Tanaka Hao 314308943@qq.com
- * @LastEditTime: 2022-11-12 21:54:32
+ * @LastEditTime: 2022-11-12 23:37:51
  * @FilePath: \C\C Primer Plus\Charpter 3\3.5.c
  * @Description: 输入年龄对应秒数
  * 
@@ -14,12 +14,12 @@
 int main(void)
 {
     int age;
-    long sec;
+    long long sec;
     printf("Enter your age:");
     scanf("%d", &age);
-    sec=(long)age*3.156e7;
+    sec=(long long)age*3.156e7;
 
-    printf("That is %ld seconds.", sec);
+    printf("That is %lld seconds.", sec);
 
     getchar();
     getchar();
